Fix tests that depend on math/rand for uniqueness.
What
===
Fix flaky tests that depend on math/rand for uniqueness by replacing the unix timestamp based random number generation with crypto random.

Why
===
The tests were using math/rand to make the test data that is created in the sandbox environment unique to that test run. Tests were failing often on TravisCI because the tests are run on multiple versions of Go at the same time. At roughtly the same time the test was running, and creating transactions that had the same customer name because the tests start at the same time resulting in the unix timestamp being the same. Using crypto/rand significantly decreases the possibility of collision.