Fix the flaky settlement summary test
What
===
Sort the results that get compared at the end of the settlement summary
test, before they are compared.

Why
===
Sometimes the results come back in different orders. The order doesn't
matter, we just want to check that the items are the same and sorting
them will ensure the tests passes even if they come back in a different
order.