Move test env vars from .travis.yml to .env
What
===
Move test env vars from .travis.yml to .env.

Why
===
To reduce the barrier to entry to make small changes, run tests and post
PRs. Ideally when someone works on the library they'd signup for a
sandbox account and copy their keys to .env, but a lot of the time it's
unnecessary for every contributor to have their own account - they just
want to make a change and check that tests still pass. I've noticed this
myself even.

Note
===
Some tests don't play well when run in parallel or not on Travis, so
independent of this change there's still work to be done to ensure tests
pass everytime and consistently.