Add descriptor to Transactions and Subscriptions
What
===
Add descriptor to `Transaction` and `Subscriptions`.

Why
===
So that a merchant can set and see the dynamic descriptor if their
account supports it.

Notes
===
The expectations for descriptors on `Transaction`s are not included in
the `TestAllTransactionFields` test because they will only pass if the
sandbox account is configured for dynamic descriptors but are instead
separated into their own test with an explanation on that tests
dependency.