Preserve the start element when building the xml
What
===
Copy the original start element and give it the new name instead of
creating a new start element.

Why
===
If we ever add attributes they'd be lost. It's mostly irrelevant now
because we don't support attributes in the model, but if attributes were
added they'd be discarding here. We generally don't need to create our
own element for the root in our object.