Rename RangeDateField to TimeField
What
===
Rename RangeDateField to TimeField.

Why
===
The field contains times, not dates. The other Braintree SDKs don't seem
to have any consensus about what these types of fields are called
because each language has a different way of implementing this. In the
Braintree docs these fields are discussed as both date and time
searches, so given that dats and times are represented as the
`time.Time` struct in go I think naming this to match makes the most
sense.