Parallelize tests
What
===
Mark all the tests that should be able to run in parallel as
tests that can be run in parallel.

Why
===
Build time is ~47 seconds, but most tests are blocking on network calls.
Running the tests in parallel across 15 routines seems like the sweet
spot where the tests will complete in ~5-6 seconds on a single core CPU.