Fix timezone in expectations on dates from Braintree
What
===
Change the expectation in the trial subscription tests that calculate
the first billing date so that the expected billing day number is based
on `US/Mountain` time, and allow others to override that with an
environment variable.

Why
===
The trial subscription tests are failing on the billing day of month for
a few hours each day and that suggests the expectation is being
calculated on a different timezone than Braintree is calculating it when
defining the billing day of month. It looks like the Braintree Java SDK
has tests like this and they use `US/Mountain` time [1] to calculate the
billing day of month. According to Braintree development docs [2] they
calculate it off the timezone of the account which in the case of the
account must be this. In the few tests I've run it seems to pass every
time I've used this timezone, so I think that is the timezone of the
test account.

Since the timezone is dependent on the timezone of the sandbox account
it makes sense to allow it to be configurable.

[1]
https://github.com/braintree/braintree_java/blob/8bde15986e9c94ce6611cdb3c7f0776f795df5a4/src/test/java/com/braintreegateway/integrationtest/SubscriptionIT.java#L141
[2]
https://developers.braintreepayments.com/guides/recurring-billing/create/ruby#subscription-days