Restructure environments for easy overriding
What
===
Change `Environment` to be a `struct` containing `BaseURL` instead of it
being a type with underlying type string and a `BaseURL()` method.

Why
===
The `BaseURL()` function only returns hardcoded URLs and cannot
be overriden. This prevents a developer using the library from changing
the BaseURL. Interest in doing this was expressed in PR #56.

Using an exported struct with exported fields makes it easy for a
developer to construct their own environment. The use of a struct also
lets us expand it with other parameters if needed without breaking
existing uses, which is important as this change makes `Environment`
part of the exported interface of the package.

Related
===
Closes #56.