Unmarshal transaction processor response code when empty
What
===
Added a type for the ProcessorResponseCode on Transaction that
unmarshals itself and considers an empty XML tag as zero.

Why
===
In some situations, e.g. a gateway reject, the processor response code
field will be an empty string. The processor response code field on
transaction is an int, but the default xml decoder will error when
unmarshaling an empty string into an int.

Why not nullable.NullInt64
===
In most places we've used `nullable.NullInt64` which is a struct that
wraps an Int64 to solve a similar problem. In the case of the processor
response code this has been a well established part of the API and
changing it from a builtin int to a struct would be more disruptive to
existing integrations than to a type with an underlying type of int.