Restructure environments for easy overriding (#114)
What
===
Change `Environment` to be a `struct` containing the base URL instead of it being a type with underlying type string and a `BaseURL()` method.

Why
===
The `BaseURL()` function only returns hardcoded URLs and cannot be overriden. This prevents a developer using the library from changing the BaseURL. Interest in doing this was expressed in PR #56.