Tighten the depencencies of hmacer
What
===
Remove dependency on Braintree struct from hmacer, replacing it with an
explicit public and private key as constructor inputs.

Why
===
There's no need for the hmacer to need the full Braintree struct as
input since it only requires a public and private key. Once we add
credential types that do not include public and private keys (e.g.
access tokens) tightening hmacer's dependencies will give us a clearer
boundary for where we'll need to check we have the credentails we need
to support hmacing.