Rename ApiVersion to apiVersion
What
===
Rename ApiVersion to apiVersion making it unexported.

Why
===
The type is only intended for internal use and no public function or
field exposes it. It also unnecessarily clutters the go docs for the
package. The type and variables were added in a recent PR but the fact
they were exported was accidental.