Fixed settlement batch summary unmarshaling.
What
===
Change the unmarshaling of the settlement batch summary records
`MerchantAccountId` field to unmarshal from the XML tag
`merchant-account_id` to `merchant-account-id`, and change the
`TestSettlementBatch` test to test the fields in the record.

The rewrite subsequently makes the `TestSettlementBatch` test more
reliable and less flaky.

Why
===
The XML tag was incorrect for the `MerchantAccountId` field so the field
was always an empty string. The test didn't test this field and was
generally flaky and would fail even when the data was correct.