Correct the test accounts timezone
What
===
Change the timezone used in the test account to `America/Chicago`.

Why
===
It was set to `US/Mountain` because that's what we saw the Braintree
Java SDK using, but the Go SDK has it's own test account in sandbox and
I double checked in the control panel what timezone that was, and it
said US Central Time, which Chicago uses.