Remove CustomerID field from Transaction
What
===
Remove `CustomerID` field from `Transaction`.

Why
===
The field is not returned in a transaction response and only exists in
the transaction request. It is an artifact of when the `Transaction`
struct was used for both requests and responses.  The field also exists
on the `TransactionRequest` struct where it should exist, but it doesn't
need to exist on `Transaction.

Ref
===
Resolves #150.