Make XML deserialization respect nil attribute
What
===
Remove any element in the XML from the Braintree API that has the
`nil="true"` attribute before the XML is deserialized into structs.

Why
===
Braintree's XML format adds the `nil="true"` attribute to elements that
are intended to be nil and not have a default value. In the official
Braintree SDKs it looks like values on things like `Transaction` are
explicitly set to `null`/`nil` when the attribute is present.

As it stands the Go SDK ignores the attribute and we end up with default
values. At this stage this only impacts one element on `Subscription`,
but there are planning to add more fields (e.g. tax_amount) that would
be impacted.