Remove support for go1.2-1.5.
What
===
Remove support for go1.2-1.5 leaving support for go1.6+.

Why
===
Supporting older versions is becoming a significant constraint. Given
that all go versions have a go1 compatibility promise and go1.6 is
one year old, it's reasonable to expect newer versions. This will make
it easier to support new features like context (#156) and default
timeouts (#133).

Ref
===
There is discussion in #158 that led to the decision to update the
compatible verions.