Reduce the many for loops in DefaultPaymentMethod
What
===
Reduce the many for loops in `Customer` `DefaultPaymentMethod()` by
getting the full list of `PaymentMethod`s from `Customer`
`PaymentMethods()` and iterating on that instead of iterating on each
payment methods slice separately.

Why
===
As we add payment methods it's one less place we run the risk of
forgetting to add a payment method here and the function not returning
the correct payment method for a customer. There's a small performance
hit to putting all the payment methods in a slice first then iterating
it, but considering this package makes network calls any performance hit
is neglible in comparison to other operations the package performs.