Remove nullables from search result structs
What
===
Remove nullables from search result structs.

Why
===
They are unnecessary. From what I can tell the structs will always have
values for the number of items in the struct. If in some case we
received an error response and the struct was built and returned it
would show a total count of zero, which is not so bad and better than
dealing with null objects.