Allow card verification to be turned off (#190)
What
===
Change the type of the `VerifyCard` option from `bool` to `*bool`.

Why
===
For merchants who have payment method verification required in their
control panel, they are unable to set the value of `VerifyCard` to
`false` and disable verification for specific payment methods like they
can with other Braintree SDKs.

This happens because the `VerifyCard` field of `CreditCardOptions` and
`PaymentMethodRequestOptions` is tagged as `omitempty`, because we only
want to send the value to the gateway if the merchant has explicitly set
a value. Combined with the type of `bool`, this means if a merchant sets
the value as `false`, we don't send it to the gateway.

Making this field a `*bool` and `omitempty` will mean the field is not
sent if it is `nil` (not set), but will be sent if it is given the value
of `false` or `true.

This issue was raised in #179.

Note
===
There is additional marshaling code added to make the VerifyCard field
marshal correctly on versions of go prior to go1.8.