Gzip decode only if Content-Encoding is gzip
What
===
Restrict Gzip decoding to only when the response indicates that the body
is Gzipped, by looking at the `Content-Encoding` and see if it is
`gzip`.

Why
===
When running tests I've started to see gzip errors every now and then.
In those cases error responses were being returned were 500, 504 and 403
errors, and these don't seem to be gzipped.

Ref
===
The Braintree Java SDK it only treats the body as gzipped if the
Content-Encoding as well:
https://github.com/braintree/braintree_java/blob/acc7c9ada19e09f6181dd501657fc7f4f1b7afcb/src/main/java/com/braintreegateway/util/Http.java#L161