Add log flag to go test runs.
What
===
Add `log` flag to go test runs. If `-log` is added to any `go test` the
tests will log the xml sent and received with the Braintree API.

Why
===
I often find myself setting a logger when inspecting messages when
adding new tests or investigating an issue. It seems like it'd be handy
to be able to turn it on/off as needed.