Exclude test_values.go from non-test builds
What
===
Exclude test_values.go from non-test builds.

Why
===
As reportd in #195 the code in `test_values.go` is included in the
behavior of the exported package. I don't think this was ever intended
as everything inside `test_values.go` is unexported and only used by the
`_test.go` test files. The fact it is included is a problem since
04a8179 #192 because an `init()` funtion was introduced into that file
in that commit, with the intention that the function only be triggered
in test runs.