Exclude test_values.go from non-test builds (#196)
What
===
Exclude test_values.go from non-test builds.

Why
===
As reportd in #195 the code in `test_values.go` is included in the behavior of the exported package. I don't think this was ever intended as everything inside `test_values.go` is unexported and only used by the `_test.go` test files. The fact it is included is a problem since 04a8179 #192 introduced an `init()` function into that file, that calls `flag.Parse`. The intention when that change was made was that the code in that file would only be built in test runs.

Renaming the file so that it's suffix is `_test.go` causes it only to be included when building for tests.