Remove omitempty from structs only used for deserialization (#201)
What
===
Remove omitempty XML tag option from struct fields only used for
deserialization.

Why
===
The omitempty option on struct fields only impacts whether the field is
serialized if it contains an empty value. When deserializing the option
has no impact, and these structs are only used to deserialize responses
from the gateway. The tags on the structs have created confusion on
what the structs are for since it makes it look like they are used for
serialization.