Add CustomerRequest (#212)
What
===
Add CustomerRequest and change the CustomerGateway Create function to use
the same patterns found in the official Braintree SDKs.

Why
===
Not all the fields that are returned for a customer are valid to pass to
Braintree when creating and updating customers. The official Braintree
SDKs, and some of the other gateways in this SDK, have moved to the
model of separating requests and responses. This is part of the
continuing effort to bring this SDK inline with the official SDKs.

Notes
===
This contains a breaking change. The object used to hold the parameters
for the CustomerGateway's Create function has changed from Customer to
CustomerRequest. This is to be consistent with the other SDKs and so
that the request struct only contains fields that can be passed to
Braintree which makes its use clearer.