Rewrite validation errors handling
Rewrite the parsing of validation errors, so that all objects and fields
returned in an error response are available.

In the process several bugs with how validation errors were parsed are
fixed. No longer do you need to call
`.For("Transaction").For("Base").For("Base")` to
get to the base errors on a Transaction, instead you call
`.For("Transaction").For("Base")`.

Since so much was changing, I also normalized some of the names of
structs with the names used in the official Braintree SDKs, so it's
easier to understand how to use the this SDK by reading the standard
Braintree developer docs.