Fix errcheck fmt.Fprintf analysis failure
What
===
Explicitly indicate that we don't care about the error returned by
`fmt.Fprintf` in the example code.

Why
===
Errcheck started considering `fmt.Fprintf` as a function that returns an
error that should be checked. We wouldn't do anything if an error was
returned by the function in the example so setting the error to `_` will
have errcheck ignore it.