Remove flaky test not offering much value (#234)
What
===
Remove the `TestMerchantAccountTransaction` test.

Why
===
The test fails 25-50% of the time. The test is creating a merchant
account and then attempting to transact on it. The fields used in the
tests are already tested in other tests, so it adds little value. I
don't know why it fails sometime, but I'm guessing since the merchant
account has just been created maybe creating a transaction immediately
is just too soon sometimes.