Add SearchPage to TransactionGateway and SubscriptionGateway (#247)
What
===
Add `SearchPage` to `TransactionGateway` and `SubscriptionGateway` that
can be used to page through search results returned by `SearchIDs`.
Deprecate other search functions, `Search` and `SearchNext`.

Why
===
The existing pagination pattern of using the `Search` and `SearchNext`
functions are not simple to program with, and they limit access to the
search space and assume the developer will want to search sequentially
through the pages.

The complexity with using the functions primarily comes from the fact
that the developer gets paged results from two different functions. This
means the developer cannot put those functions merely in a for loop and
iterate. They must call one first, deal with those results, then loop on
the second function. It is less intuitive to write code this way, and
much easier to think about if we get the IDs first, then just iterate
over them in pages.