Reduce subscription integration test times
What
===
Change the period of data that is downloaded during the Subscription
pagination tests so it looks only at the last 10 minutes and not the
last 24 hours.

Why
===
When the tests have been run many times in a day the number of
subscription builds up and these integration tests are pulling almost
10,000 subscriptions down across 200 pages, which means there's 200
requests and responses which was taking ~15 minutes for a single test.
This test hadn't changed recently but the tests were ran a lot in the
last few days and that caused the subscription count to sky rocket. The
10 minute window should keep the number low enough.