Remove the winning and losing dispute tests (#261)
### What
Remove the winning and losing dispute tests.

### Why
The two tests hardly pass as the state still comes back as `disputed` for a period of time after we submit the winning or losing evidence. We don't really need the tests because they really just test Braintree's sandbox environment. The other tests we have for the dispute gateway test the same functions.