Add more fields and options to client token generate (#265)
What
===
Add the full range of fields that are available in the official SDKs on the ClientTokenGateway's Generate function call.

Why
===
This aligns the functionality available on client token generation with the [official libraries](https://developers.braintreepayments.com/reference/request/client-token/generate).

This change support all documented parameters but also fix certain issues for some users. In particular, when the user wants to use `ApplePay` or `GooglePlay` and has more than one merchant account (E.g. sub accounts per country), the client token generation defaults to the main account if no `merchantAccountId` is given. This causes issues on the client SDK (browser or mobile) if, for example, `ApplePay` or `GooglePlay` is not enabled in the country for the default account.