Fix XML marshaling of ClientTokenRequestOptions for Go 1.7 (#268)
What
===
Don't serialize the ClientTokenRequestOptions VerifyCard *bool field if it is false in Go 1.7.

Why
===
The VerifyCard *bool field was just recently added in abe29b19ee6b20d619a49772d73b1e8427c6ea47 and it relies on the default serialization logic for fields using tags. This is fine in Go 1.8+ but in versions of Go prior to that there was a bug serializing *bool fields. Tests were added to demonstrate how the master branch was in a broken state in f0f9ed064169164c4ab83a453fdefc3764bde178.