Serialize decimals to string consistently (#280)
What
===
Change the number of zeros added to the beginning of a decimal that will
only render non-zero digits below a decimal point so that there is
always one zero before the decimal point.

Why
===
The decimal serialization to a string appears to have a couple of
oddities.

`NewDecimal(55, 2)` serialized to `".55"`.
`NewDecimal(55, 3)` serialized to `"00.055"`.

When calculating the number of zeros to add to the beginning of a number
if the scale is larger than the number, it wasn't taking into account if
the length was equal to the scale (e.g. 0.55 where there are two digits
provided and a scale of 2), and the other situation above. The number of
zeros being added was the scale and not the number of zeros missing to
consistently generate a string leading with `0.`.